Use decimal for the san balance

+++ priv/repo/structure.sql
@@ -989,7 +989,7 @@ CREATE TABLE public.user_api_key_tokens (
     inserted_at timestamp without time zone NOT NULL,
     updated_at timestamp without time zone NOT NULL,
     san_tokens integer,
-    san_balance integer,
+    san_balance numeric,
     san_balance_updated_at timestamp without time zone
 );
